Clean up code XML docs and break up enum classes into their own files.
