+++ toast_ui.blazor_calendar/toast_ui.blazor_calendar.csproj
@@ -46,12 +46,12 @@
     <Exec Command="npm install" WorkingDirectory="NpmJS" />
     <Exec Command="npm run dev" WorkingDirectory="NpmJS" />
   </Target>
-<!--
+
   <Target Name="PreBuildA" BeforeTargets="PreBuildEvent" Condition="'$(Configuration)' == 'Release'">
     <Exec Command="npm install" WorkingDirectory="NpmJS" />
     <Exec Command="npm run wbp" WorkingDirectory="NpmJS" />
   </Target>
-  -->
+
 <!--
   <Target Name="PreBuild1" BeforeTargets="PreBuildEvent">
     <Exec Command="npm install" WorkingDirectory="NpmJS" />
